Fix contact ident & responsible person saving/diaplay
