+++ app/models/concerns/audit.rb
@@ -1,9 +1,36 @@
 module Audit
   extend ActiveSupport::Concern
-  include AuditBase
 
   included do
     attr_accessor :version_loader
+
+    before_create :add_creator
+    before_create :add_updator
+    before_update :add_updator
+
+    def add_creator
+      self.creator_str = User.whodunnit || 'console-root'
+      true
+    end
+
+    def add_updator
+      self.updator_str = User.whodunnit || 'console-root'
+      true
+    end
+
+    def creator
+      return if creator_str.blank?
+
+      creator = user_from_id_role_username creator_str
+      creator.presence || creator_str
+    end
+
+    def updator
+      return if updator_str.blank?
+
+      updator = user_from_id_role_username updator_str
+      updator.presence || updator_str
+    end
 
     def user_from_id_role_username(str)
       registrar = Registrar.find_by(name: str)
